requirements.txt
Django>=3.0,<4.0
psycopg2==2.9.3
djangorestframework==3.13.1
drf-nested-routers==0.93.4